+++ internal/handlers/mac_bundle_xattr_other.go
@@ -1,7 +1,0 @@
-//go:build !darwin
-
-package handlers
-
-func clearMacAppQuarantine(string) error {
-	return nil
-}